+++ include_headers/decoder.h
@@ -3,6 +3,6 @@
 
 # include "header.h"
 
-unsigned char *decode_msg(t_daniel *daniel_body, unsigned char *decompressed, short len);
+unsigned char *decode_msg(t_daniel *daniel_body, unsigned char *decompressed, size_t len);
 
 #endif
